refactor: AutoDto 를 AutoCommand 로 수정
