+++ backend/footprint/src/main/java/foot/footprint/domain/member/dto/auth/LoginCommand.java
@@ -1,6 +1,5 @@
-package foot.footprint.domain.member.dto.authDto;
+package foot.footprint.domain.member.dto.auth;
 
-import foot.footprint.domain.member.dto.authRequest.LoginRequest;
 import lombok.Getter;
 
 @Getter
