+++ backend/footprint/src/main/java/foot/footprint/domain/member/dto/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package foot.footprint.domain.member.dto.authRequest;
+package foot.footprint.domain.member.dto.auth;
 
 import lombok.Getter;
