+++ backend/footprint/src/main/java/foot/footprint/domain/member/dto/auth/SignUpRequest.java
@@ -1,4 +1,4 @@
-package foot.footprint.domain.member.dto.authRequest;
+package foot.footprint.domain.member.dto.auth;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
