+++ backend/footprint/src/test/java/foot/footprint/service/member/AuthServiceTest.java
@@ -6,8 +6,8 @@ import foot.footprint.domain.member.dao.MemberRepository;
 import foot.footprint.domain.member.domain.AuthProvider;
 import foot.footprint.domain.member.domain.Role;
 import foot.footprint.domain.member.domain.Member;
-import foot.footprint.domain.member.dto.authDto.LoginCommand;
-import foot.footprint.domain.member.dto.authDto.SignUpCommand;
+import foot.footprint.domain.member.dto.auth.LoginCommand;
+import foot.footprint.domain.member.dto.auth.SignUpCommand;
 import foot.footprint.domain.member.exception.AlreadyExistedEmailException;
 import foot.footprint.domain.member.exception.NotMatchPasswordException;
 import foot.footprint.global.error.exception.NotExistsException;
